+++ sdk/network/arm-network/src/operationsInterfaces/firewallPolicyRuleCollectionGroups.ts
@@ -15,7 +15,14 @@ import {
   FirewallPolicyRuleCollectionGroupsGetOptionalParams,
   FirewallPolicyRuleCollectionGroupsGetResponse,
   FirewallPolicyRuleCollectionGroupsCreateOrUpdateOptionalParams,
-  FirewallPolicyRuleCollectionGroupsCreateOrUpdateResponse
+  FirewallPolicyRuleCollectionGroupsCreateOrUpdateResponse,
+  FirewallPolicyDraftRuleCollectionGroup,
+  FirewallPolicyRuleCollectionGroupsCreateOrUpdateDraftOptionalParams,
+  FirewallPolicyRuleCollectionGroupsCreateOrUpdateDraftResponse,
+  FirewallPolicyRuleCollectionGroupsGetDraftOptionalParams,
+  FirewallPolicyRuleCollectionGroupsGetDraftResponse,
+  FirewallPolicyRuleCollectionGroupsDeleteDraftOptionalParams,
+  FirewallPolicyRuleCollectionGroupsDeleteDraftResponse
 } from "../models";
 
 /// <reference lib="esnext.asynciterable" />
@@ -110,4 +117,52 @@ export interface FirewallPolicyRuleCollectionGroups {
     parameters: FirewallPolicyRuleCollectionGroup,
     options?: FirewallPolicyRuleCollectionGroupsCreateOrUpdateOptionalParams
   ): Promise<FirewallPolicyRuleCollectionGroupsCreateOrUpdateResponse>;
+  /**
+   * Creates or updates a new draft version of the specified Firewall Policy Rule Collection Group.
+   * @param resourceGroupName The name of the resource group.
+   * @param firewallPolicyName The name of the Firewall Policy.
+   * @param ruleCollectionGroupName The name of the FirewallPolicyRuleCollectionGroup.
+   * @param parameters Parameters supplied to the create or update FirewallPolicyRuleCollectionGroup
+   *                   operation.
+   * @param options The options parameters.
+   */
+  createOrUpdateDraft(
+    resourceGroupName: string,
+    firewallPolicyName: string,
+    ruleCollectionGroupName: string,
+    parameters: FirewallPolicyDraftRuleCollectionGroup,
+    options?: FirewallPolicyRuleCollectionGroupsCreateOrUpdateDraftOptionalParams
+  ): Promise<FirewallPolicyRuleCollectionGroupsCreateOrUpdateDraftResponse>;
+  /**
+   * Get the current draft version of the specified Firewall Policy Rule Collection Group.
+   * @param resourceGroupName The name of the resource group.
+   * @param firewallPolicyName The name of the Firewall Policy.
+   * @param ruleCollectionGroupName The name of the FirewallPolicyRuleCollectionGroup.
+   * @param parameters Parameters supplied to the create or update FirewallPolicyRuleCollectionGroup
+   *                   operation.
+   * @param options The options parameters.
+   */
+  getDraft(
+    resourceGroupName: string,
+    firewallPolicyName: string,
+    ruleCollectionGroupName: string,
+    parameters: FirewallPolicyRuleCollectionGroup,
+    options?: FirewallPolicyRuleCollectionGroupsGetDraftOptionalParams
+  ): Promise<FirewallPolicyRuleCollectionGroupsGetDraftResponse>;
+  /**
+   * Get the current draft version of the specified Firewall Policy Rule Collection Group.
+   * @param resourceGroupName The name of the resource group.
+   * @param firewallPolicyName The name of the Firewall Policy.
+   * @param ruleCollectionGroupName The name of the FirewallPolicyRuleCollectionGroup.
+   * @param parameters Parameters supplied to the create or update FirewallPolicyRuleCollectionGroup
+   *                   operation.
+   * @param options The options parameters.
+   */
+  deleteDraft(
+    resourceGroupName: string,
+    firewallPolicyName: string,
+    ruleCollectionGroupName: string,
+    parameters: FirewallPolicyRuleCollectionGroup,
+    options?: FirewallPolicyRuleCollectionGroupsDeleteDraftOptionalParams
+  ): Promise<FirewallPolicyRuleCollectionGroupsDeleteDraftResponse>;
 }