+++ sdk/network/arm-network/CHANGELOG.md
@@ -1,21 +1,76 @@
 # Release History
     
-## 30.1.0 (2022-12-13)
+## 31.0.0 (2022-12-14)
     
 **Features**
 
+  - Added operation FirewallPolicies.beginDeployDraft
+  - Added operation FirewallPolicies.beginDeployDraftAndWait
+  - Added operation FirewallPolicies.createOrUpdateDraft
+  - Added operation FirewallPolicies.deleteDraft
+  - Added operation FirewallPolicies.getDraft
+  - Added operation FirewallPolicyRuleCollectionGroups.createOrUpdateDraft
+  - Added operation FirewallPolicyRuleCollectionGroups.deleteDraft
+  - Added operation FirewallPolicyRuleCollectionGroups.getDraft
   - Added Interface DdosProtectionPlansDeleteHeaders
   - Added Interface DelegationProperties
+  - Added Interface FirewallPoliciesCreateOrUpdateDraftOptionalParams
+  - Added Interface FirewallPoliciesDeleteDraftOptionalParams
+  - Added Interface FirewallPoliciesDeployDraftOptionalParams
+  - Added Interface FirewallPoliciesGetDraftOptionalParams
+  - Added Interface FirewallPolicyDraft
+  - Added Interface FirewallPolicyDraftRuleCollectionGroup
+  - Added Interface FirewallPolicyRuleCollectionGroupsCreateOrUpdateDraftOptionalParams
+  - Added Interface FirewallPolicyRuleCollectionGroupsDeleteDraftOptionalParams
+  - Added Interface FirewallPolicyRuleCollectionGroupsGetDraftOptionalParams
   - Added Interface PartnerManagedResourceProperties
   - Added Interface PublicIPAddressesDdosProtectionStatusHeaders
   - Added Interface PublicIPAddressesDeleteHeaders
+  - Added Type Alias FirewallPoliciesCreateOrUpdateDraftResponse
+  - Added Type Alias FirewallPoliciesGetDraftResponse
+  - Added Type Alias FirewallPolicyRuleCollectionGroupsCreateOrUpdateDraftResponse
+  - Added Type Alias FirewallPolicyRuleCollectionGroupsDeleteDraftResponse
+  - Added Type Alias FirewallPolicyRuleCollectionGroupsGetDraftResponse
   - Interface BackendAddressPool has a new optional parameter virtualNetwork
+  - Interface DdosProtectionPlan has a new optional parameter publicIPAddresses
   - Interface NetworkVirtualAppliance has a new optional parameter delegation
   - Interface NetworkVirtualAppliance has a new optional parameter deploymentType
   - Interface NetworkVirtualAppliance has a new optional parameter partnerManagedResource
   - Interface PolicySettings has a new optional parameter customBlockResponseBody
   - Interface PolicySettings has a new optional parameter customBlockResponseStatusCode
+  - Interface VirtualNetwork has a new optional parameter flowLogs
   - Added function getContinuationToken
+
+**Breaking Changes**
+
+  - Interface AdminRuleCollectionsListNextOptionalParams no longer has parameter skipToken
+  - Interface AdminRuleCollectionsListNextOptionalParams no longer has parameter top
+  - Interface AdminRulesListNextOptionalParams no longer has parameter skipToken
+  - Interface AdminRulesListNextOptionalParams no longer has parameter top
+  - Interface ConnectivityConfigurationsListNextOptionalParams no longer has parameter skipToken
+  - Interface ConnectivityConfigurationsListNextOptionalParams no longer has parameter top
+  - Interface DdosProtectionPlan no longer has parameter publicIpAddresses
+  - Interface ManagementGroupNetworkManagerConnectionsListNextOptionalParams no longer has parameter skipToken
+  - Interface ManagementGroupNetworkManagerConnectionsListNextOptionalParams no longer has parameter top
+  - Interface NetworkGroupsListNextOptionalParams no longer has parameter skipToken
+  - Interface NetworkGroupsListNextOptionalParams no longer has parameter top
+  - Interface NetworkInterfacesListVirtualMachineScaleSetIpConfigurationsNextOptionalParams no longer has parameter expand
+  - Interface NetworkManagersListBySubscriptionNextOptionalParams no longer has parameter skipToken
+  - Interface NetworkManagersListBySubscriptionNextOptionalParams no longer has parameter top
+  - Interface NetworkManagersListNextOptionalParams no longer has parameter skipToken
+  - Interface NetworkManagersListNextOptionalParams no longer has parameter top
+  - Interface ScopeConnectionsListNextOptionalParams no longer has parameter skipToken
+  - Interface ScopeConnectionsListNextOptionalParams no longer has parameter top
+  - Interface SecurityAdminConfigurationsListNextOptionalParams no longer has parameter skipToken
+  - Interface SecurityAdminConfigurationsListNextOptionalParams no longer has parameter top
+  - Interface ServiceTagInformationListNextOptionalParams no longer has parameter noAddressPrefixes
+  - Interface ServiceTagInformationListNextOptionalParams no longer has parameter tagName
+  - Interface StaticMembersListNextOptionalParams no longer has parameter skipToken
+  - Interface StaticMembersListNextOptionalParams no longer has parameter top
+  - Interface SubscriptionNetworkManagerConnectionsListNextOptionalParams no longer has parameter skipToken
+  - Interface SubscriptionNetworkManagerConnectionsListNextOptionalParams no longer has parameter top
+  - Interface VirtualNetworksListDdosProtectionStatusNextOptionalParams no longer has parameter skipToken
+  - Interface VirtualNetworksListDdosProtectionStatusNextOptionalParams no longer has parameter top
     
     
 ## 30.0.0 (2022-09-29)